+++ example/android/app/src/main/kotlin/com/pi/jsondynamicwidget/example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.pi.jsondynamicwidget.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}