+++ lib/src/schema/schemas/merge_semantics_schema.dart
@@ -1,0 +1,13 @@
+class MergeSemanticsSchema {
+  static const id =
+      'https://peiffer-innovations.github.io/flutter_json_schemas/schemas/json_dynamic_widget/merge_semantics.json';
+
+  static final schema = {
+    r'$schema': 'http://json-schema.org/draft-06/schema#',
+    r'$id': '$id',
+    r'$comment':
+        'https://api.flutter.dev/flutter/widgets/MergeSemantics-class.html',
+    'type': 'null',
+    'title': 'MergeSemanticsBuilder',
+  };
+}